+++ app/src/androidTest/java/woowacourse/movie/view/activity/MovieReservationActivityTest.kt
@@ -17,9 +17,9 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
 import woowacourse.movie.constants.MovieContentKey
-import woowacourse.movie.model.Date
-import woowacourse.movie.model.MovieContent
 import woowacourse.movie.model.data.MovieContentsImpl
+import woowacourse.movie.model.data.dto.Date
+import woowacourse.movie.model.data.dto.MovieContent
 
 @RunWith(AndroidJUnit4::class)
 class MovieReservationActivityTest {
