+++ app/src/main/java/woowacourse/movie/conrtract/MovieReservationContract.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.conrtract
 
-import woowacourse.movie.model.MovieContent
+import woowacourse.movie.model.data.dto.MovieContent
 import woowacourse.movie.view.BaseView
 
 interface MovieReservationContract {
