+++ app/src/main/java/woowacourse/movie/model/data/dto/Date.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.data.dto
 
 class Date(
     val year: Int,
