+++ app/src/main/java/woowacourse/movie/model/data/dto/MovieContent.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.data.dto
 
 data class MovieContent(
     val imageId: Int,
