+++ app/src/main/java/woowacourse/movie/presenter/MovieContentListPresenter.kt
@@ -1,8 +1,8 @@
 package woowacourse.movie.presenter
 
 import woowacourse.movie.conrtract.MovieContentListContract
-import woowacourse.movie.model.MovieContent
 import woowacourse.movie.model.data.MovieContentsImpl
+import woowacourse.movie.model.data.dto.MovieContent
 
 class MovieContentListPresenter(private val view: MovieContentListContract.View) :
     MovieContentListContract.Presenter {
